+++ lib/src/sub_scope_manager.dart
@@ -1,0 +1,11 @@
+// @license
+// Copyright (c) 2019 - 2024 Dr. Gabriel Gatzsche. All Rights Reserved.
+//
+// Use of this source code is governed by terms that can be
+// found in the LICENSE file in the root of this package.
+
+/// Creates or removes sub-scopes depending on components provided by suppliers
+class SubScopeManager {
+  /// Constructor
+  const SubScopeManager();
+}