+++ lib/supply_chain.dart
@@ -15,4 +15,5 @@ export 'src/scm.dart';
 export 'src/scope.dart';
 export 'src/sub_scope_manager_blue_print.dart';
 export 'src/scope_blue_print.dart';
+export 'src/sub_scope_manager.dart';
 export 'src/tools.dart';